Refactor radio subclass into one file
